@@ -16,6 +16,7 @@ import javax.swing.JFrame;
 import javax.swing.JOptionPane;
 import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.parsers.SAXParserFactory;
+
 import org.apache.log4j.Logger;
 import org.xml.sax.InputSource;
 import org.xml.sax.SAXException;
@@ -28,12 +29,15 @@ import org.xml.sax.XMLReader;
  * @version 1.0
  */
 class LoadXMLAction
-    extends AbstractAction
-{
-    /** use to log messages **/
+        extends AbstractAction {
+    /**
+     * use to log messages
+     **/
     private static final Logger LOG = Logger.getLogger(LoadXMLAction.class);
 
-    /** the parent frame **/
+    /**
+     * the parent frame
+     **/
     private final JFrame mParent;
 
     /**
@@ -41,14 +45,19 @@ class LoadXMLAction
      * single file.
      */
     private final JFileChooser mChooser = new JFileChooser();
+
     {
         mChooser.setMultiSelectionEnabled(false);
         mChooser.setFileSelectionMode(JFileChooser.FILES_ONLY);
     }
 
-    /** parser to read XML files **/
+    /**
+     * parser to read XML files
+     **/
     private final XMLReader mParser;
-    /** the content handler **/
+    /**
+     * the content handler
+     **/
     private final XMLFileHandler mHandler;
 
 
@@ -56,13 +65,12 @@ class LoadXMLAction
      * Creates a new <code>LoadXMLAction</code> instance.
      *
      * @param aParent the parent frame
-     * @param aModel the model to add events to
-     * @exception SAXException if an error occurs
+     * @param aModel  the model to add events to
+     * @throws SAXException                 if an error occurs
      * @throws ParserConfigurationException if an error occurs
      */
     LoadXMLAction(JFrame aParent, MyTableModel aModel)
-        throws SAXException, ParserConfigurationException
-    {
+            throws SAXException, ParserConfigurationException {
         mParent = aParent;
         mHandler = new XMLFileHandler(aModel);
         mParser = SAXParserFactory.newInstance().newSAXParser().getXMLReader();
@@ -71,6 +79,7 @@ class LoadXMLAction
 
     /**
      * Prompts the user for a file to load events from.
+     *
      * @param aIgnore an <code>ActionEvent</code> value
      */
     public void actionPerformed(ActionEvent aIgnore) {
@@ -82,17 +91,17 @@ class LoadXMLAction
             try {
                 final int num = loadFile(chosen.getAbsolutePath());
                 JOptionPane.showMessageDialog(
-                    mParent,
-                    "Loaded " + num + " events.",
-                    "CHAINSAW",
-                    JOptionPane.INFORMATION_MESSAGE);
+                        mParent,
+                        "Loaded " + num + " events.",
+                        "CHAINSAW",
+                        JOptionPane.INFORMATION_MESSAGE);
             } catch (Exception e) {
                 LOG.warn("caught an exception loading the file", e);
                 JOptionPane.showMessageDialog(
-                    mParent,
-                    "Error parsing file - " + e.getMessage(),
-                    "CHAINSAW",
-                    JOptionPane.ERROR_MESSAGE);
+                        mParent,
+                        "Error parsing file - " + e.getMessage(),
+                        "CHAINSAW",
+                        JOptionPane.ERROR_MESSAGE);
             }
         }
     }
@@ -103,25 +112,20 @@ class LoadXMLAction
      * @param aFile the file to extract events from
      * @return the number of events loaded
      * @throws SAXException if an error occurs
-     * @throws IOException if an error occurs
+     * @throws IOException  if an error occurs
      */
-    private int loadFile(String aFile)
-        throws SAXException, IOException
-    {
+    public int loadFile(String aFile) throws SAXException, IOException {
         synchronized (mParser) {
             // Create a dummy document to parse the file
-            final StringBuffer buf = new StringBuffer();
-            buf.append("<?xml version=\"1.0\" standalone=\"yes\"?>\n");
-            buf.append("<!DOCTYPE log4j:eventSet ");
-            buf.append("[<!ENTITY data SYSTEM \"file:///");
-            buf.append(aFile);
-            buf.append("\">]>\n");
-            buf.append("<log4j:eventSet xmlns:log4j=\"Claira\">\n");
-            buf.append("&data;\n");
-            buf.append("</log4j:eventSet>\n");
-
-            final InputSource is =
-                new InputSource(new StringReader(buf.toString()));
+            String buf = "<?xml version=\"1.0\" standalone=\"yes\"?>\n" +
+                    "<!DOCTYPE log4j:eventSet " +
+                    "[<!ENTITY data SYSTEM \"file:///" +
+                    aFile +
+                    "\">]>\n" +
+                    "<log4j:eventSet xmlns:log4j=\"Claira\">\n" +
+                    "&data;\n" +
+                    "</log4j:eventSet>\n";
+            final InputSource is = new InputSource(new StringReader(buf));
             mParser.parse(is);
             return mHandler.getNumEvents();
         }
